// src/Types/IResponse.ts
export interface IResponse<T = string> {
    Body: T;
    Headers: {
        'accept-patch'?: string | number;
        'accept-ranges'?: string | number;
        'access-control-allow-credentials'?: string | number;
        'access-control-allow-headers'?: string | number;
        'access-control-allow-methods'?: string | number;
        'access-control-allow-origin'?: string | number;
        'access-control-expose-headers'?: string | number;
        'access-control-max-age'?: string | number;
        age?: string | number;
        allow?: string | number;
        'alt-svc'?: string | number;
        'cache-control'?: string | number;
        connection?: string | number;
        'content-disposition'?: string | number;
        'content-encoding'?: string | number;
        'content-language'?: string | number;
        'content-length'?: string | number;
        'content-location'?: string | number;
        'content-range'?: string | number;
        'content-security-policy'?: string | number;
        'content-type'?: string | number;
        date?: string | number;
        'delta-base'?: string | number;
        etag?: string | number;
        expires?: string | number;
        im?: string | number;
        'last-modified'?: string | number;
        link?: string | number;
        location?: string | number;
        'max-age'?: string | number;
        'no-cache'?: string | number;
        'no-store'?: string | number;
        pragma?: string | number;
        'proxy-authenticate'?: string | number;
        'public-key-pins'?: string | number;
        refresh?: string | number;
        'retry-after'?: string | number;
        server?: string | number;
        'set-cookie'?: string | number;
        'strict-transport-security'?: string | number;
        tk?: string | number;
        trailer?: string | number;
        'transfer-encoding'?: string | number;
        upgrade?: string | number;
        vary?: string | number;
        via?: string | number;
        warning?: string | number;
        'www-authenticate'?: string | number;
        'x-powered-by'?: string | number;
        'x-request-id'?: string | number;
        'x-ua-compatible'?: string | number;
        'x-xss-protection'?: string | number;

        [Name: string]: string | number | undefined;
    };
    StatusCode: number;
}